Fixed klassenanlage autocomplete
